requirements.txt
setuptools==70.0.0
langchain_community==0.2.6
langchain_core==0.2.10
langchain_groq==0.1.5
langchain_openai==0.1.10
newspaper3k==0.2.8
pandas==2.2.2
python-dotenv==1.0.1
selenium==4.22.0
streamlit==1.35.0
lxml==5.2.2
lxml_html_clean==0.1.1
webdriver-manager==4.0.1
streamlit-lottie==0.0.5
deep-translator==1.11.4
goose3==3.1.19